requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.13.3
packaging==24.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
redis==5.0.5
rq==1.16.2
setuptools==68.2.2
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
wheel==0.41.2
